api/requirements.txt
black==21.9b0
isort==5.9.3
flake8==4.0.1
fastapi==0.109.0
uvicorn==0.26.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
fastapi_sqlalchemy==0.2.1
alembic==1.13.1
python-dotenv==1.0.0
fastapi_sqlalchemy==0.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory_boy==3.3.0
pytest_factoryboy==2.6.0
pytest-env==1.1.3
sqlalchemy_utils==0.41.1
coverage==7.4.0
PyJWT==2.8.0
